Sigma can't be sized
For example, consider a function of type Pi (A B : Ui). Sigma (x :
A). B -> C. This function can take both (tensor i32 i32) and (tensor
i8 i8) as its arguments. Since the function must be able to access the
content of the argument in the consistent way, these products must be
aligned just in the same way. This implies that we can't "peek" the
contents of a product and try to reduce its size.